pyproject.toml
[tool.poetry]
name = "fast-api-graphql"
version = "0.1.0"
description = ""
authors = ["nkmrkz <kazutaka.nakamura@ncdc.co.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
strawberry-graphql = "^0.114.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
